+++ src/main/java/org/jabref/gui/externalfiles/UnlinkedFilesDialogViewModel.java
@@ -42,6 +42,7 @@ import org.jabref.logic.util.BackgroundTask;
 import org.jabref.logic.util.StandardFileType;
 import org.jabref.logic.util.TaskExecutor;
 import org.jabref.model.database.BibDatabaseContext;
+import org.jabref.model.entry.LinkedFile;
 import org.jabref.model.util.FileUpdateMonitor;
 
 import de.saxsys.mvvmfx.utils.validation.FunctionBasedValidator;
@@ -144,18 +145,21 @@ public class UnlinkedFilesDialogViewModel {
     }
 
     public void startImport() {
-        List<Path> fileList = checkedFileListProperty
+        List<Path> fileList = new java.util.ArrayList<>(checkedFileListProperty
                 .stream()
                 .map(TreeItem::getValue)
                 .map(FileNodeViewModel::getPath)
                 .filter(Files::isRegularFile)
-                .toList();
+                .toList());
 
         if (fileList.isEmpty()) {
             LOGGER.warn("There are no valid files checked for import");
             return;
         }
         resultList.clear();
+
+        // Removes duplicate files before importing
+        fileList.removeIf(path -> bibDatabase.getDatabase().containsEntryWithLinkedFile(new LinkedFile(path)));
 
         importFilesBackgroundTask = importHandler
                 .importFilesInBackground(fileList, bibDatabase, preferences.getFilePreferences(), TransferMode.LINK)